requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg==3.1.18
python-dotenv==1.0.1
gunicorn==21.2.0

